LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/unserializable-expected.txt



PASS updateCurrentEntry() must throw if state is unserializable (WritableStream)
FAIL updateCurrentEntry() must throw if state is unserializable (SharedArrayBuffer) assert_throws_dom: function "() => {
      iframe.contentWindow.navigation.updateCurrentEntry({ state: buffer });
    }" did not throw
